Harmonize Cell Types Across Datasets

This function integrates and harmonizes cell type annotations across multiple datasets by applying predefined unification maps and cell type labels. It uses a combination of transferred annotations and predefined maps to produce a consensus on cell type identities.

cell_type_ensembl_harmonised(
  input_read_RNA_assay,
  annotation_label_transfer_tbl = NULL,
  celltype_unification_maps = NULL,
  nonimmune = NULL,
  available_maps = c("azimuth", "blueprint", "monaco", "cellxgene")
)

Arguments

input_read_RNA_assay

SingleCellExperiment or Seurat object containing RNA assay data.

annotation_label_transfer_tbl

A tibble with annotation label transfer data.

celltype_unification_maps

A list containing mapping data frames for different sources (e.g., Azimuth, Blueprint, Monaco, and cellxgene). Default is NULL. If NULL, it retrieves default maps stored in HPCell.

nonimmune

A character vector specifying non-immune cell types. Default is NULL. If NULL, it retrieves default non-immune types from HPCell.

available_maps

A character vector of cell type annotation sources to include in the ensemble annotation process. Supported values include "azimuth", "blueprint", "monaco", and "cellxgene". By default, it uses all of the annotations.

Value

A tibble of the input SummarizedExperiment metadata enriched with unified cell type annotations and additional classification details.