Main Function for HPCell Map Test Differential Abundance

This function prepares and runs a differential abundance test pipeline using the 'targets' package. It sets up necessary files, appends scripts, and executes the pipeline.

map2_test_differential_abundance_hpc(
  data_list,
  formula_list,
  .abundance = NULL,
  store = tempfile(tmpdir = "."),
  computing_resources = crew_controller_local(workers = 1),
  cpus_per_task = 1,
  debug_job_id = NULL,
  append = FALSE,
  ...
)

Arguments

data_list

list of dataframes to be processed

formula_list

List of formula for the differential abundance test.

.abundance

(optional) A symbol or string indicating the column name in the SingleCellExperiment object to be used for abundance measures. If not explicitly provided, the function attempts to automatically detect an appropriate column by examining the first object in data_list.

store

File path for temporary storage.

computing_resources

Computing resources configuration.

cpus_per_task

Number of CPUs allocated per task.

debug_job_id

Optional job ID for debugging.

append

Flag to append to existing script.

...

additional arguments

Value

A targets pipeline output, typically a nested tibble with differential abundance estimates.