Add Dispersion Estimates to SingleCellExperiment Object

map_add_dispersion_to_se function adds dispersion estimates to each feature (gene) in a SingleCellExperiment object. Dispersion estimates are added based on the abundance measure specified.

map_add_dispersion_to_se(se_df, .col, formula = NULL, abundance = NULL)

Arguments

se_df

A data frame or list containing SingleCellExperiment objects.

.col

A symbol indicating the column in se_df that contains SingleCellExperiment objects.

formula

(Optional) A formula argument accepted for interface compatibility; not used internally.

abundance

(Optional) A character vector specifying the name of the assay to be used for dispersion estimation. If NULL or not provided, the first assay is used.

Value

The input data frame or list (se_df) with the specified .col modified to include dispersion estimates in each SingleCellExperiment object.

Details

The function iterates over each SingleCellExperiment object in the specified column of the input data frame or list. It calculates dispersion estimates for the features (genes) based on the specified abundance assay. The results are joined back to each SingleCellExperiment object. If no abundance assay is specified, the function defaults to the first assay in each SingleCellExperiment object.

Examples

if (FALSE) { # \dontrun{
# Assuming `se_list` is a list of SingleCellExperiment objects
result <- map_add_dispersion_to_se(se_list, .col = se_objects, abundance = "counts")
} # }