Detection of metacells with the SuperCell approach

This function detects metacells (former super-cells) from single-cell gene expression matrix

postprocess_SCimplify(
  preprocessed,
  cell.annotation = NULL,
  cell.split.condition = NULL,
  gamma,
  block.size = 10000,
  igraph.clustering = c("walktrap", "louvain"),
  return.singlecell.NW = TRUE,
  return.hierarchical.structure = TRUE,
  ...
)

Arguments

preprocessed

A list returned by preprocess_SCimplify containing preprocessed single-cell data, PCA results, and kNN graph.

cell.annotation

a vector of cell type annotation, if provided, metacells that contain single cells of different cell type annotation will be split in multiple pure metacell (may result in slightly larger numbe of metacells than expected with a given gamma)

cell.split.condition

a vector of cell conditions that must not be mixed in one metacell. If provided, metacells will be split in condition-pure metacell (may result in significantly(!) larger number of metacells than expected)

gamma

graining level of data (proportion of number of single cells in the initial dataset to the number of metacells in the final dataset)

block.size

number of cells to map to the nearest metacell at the time (for approx coarse-graining)

igraph.clustering

clustering method to identify metacells (available methods "walktrap" (default) and "louvain" (not recommended, gamma is ignored)).

return.singlecell.NW

whether return single-cell network (which consists of approx.N if "do.approx" or all cells otherwise)

return.hierarchical.structure

whether return hierarchical structure of metacell

...

other parameters of build_knn_graph function

Value

A tibble with column 'cell' and 'membership' indicating which metacell cluster each cell belongs to.