Preprocess metacells with the SuperCell approach

This function preprocesses a single-cell gene expression matrix for downstream simplification using PCA and k-nearest neighbor (kNN) graph construction. It includes options for scaling, feature selection, approximate sampling, and PCA computation methods.

preprocess_SCimplify(
  input_read_RNA_assay,
  assay = NULL,
  genes.use = NULL,
  genes.exclude = NULL,
  n.var.genes = min(1000, nrow(input_read_RNA_assay)),
  k.knn = 5,
  do.scale = TRUE,
  n.pc = 10,
  fast.pca = TRUE,
  do.approx = FALSE,
  approx.N = 5000,
  seed = 12345,
  ...
)

Arguments

input_read_RNA_assay

A SingleCellExperiment or Seurat object containing RNA assay data.

assay

assay used, default = "RNA"

genes.use

a vector of genes used to compute PCA

genes.exclude

a vector of genes to be excluded when computing PCA

n.var.genes

if "genes.use" is not provided, "n.var.genes" genes with the largest variation are used

k.knn

parameter to compute single-cell kNN network

do.scale

whether to scale gene expression matrix when computing PCA

n.pc

number of principal components to use for construction of single-cell kNN network

fast.pca

use irlba as a faster version of prcomp (one used in Seurat package)

do.approx

compute approximate kNN in case of a large dataset (>50'000)

approx.N

number of cells to subsample for an approximate approach. By default, 5000 cells are used for approximation to capture biological meaningful result.

seed

seed to use to subsample cells for an approximate approach

...

other parameters of build_knn_graph function

Value

A list of variables to be passed to the SuperCell::SCimplify gamma involved function.