reference_annotation_to_consensus

This function takes cell type annotations from multiple datasets (Azimuth, Monaco, Blueprint) and harmonizes them into a consensus annotation. The function utilizes predefined mappings between cell type labels in these datasets to generate standardized cell types across references.

reference_annotation_to_consensus(azimuth_input, monaco_input, blueprint_input)

Arguments

azimuth_input

A vector of cell type annotations from the Azimuth dataset.

monaco_input

A vector of cell type annotations from the Monaco dataset.

blueprint_input

A vector of cell type annotations from the Blueprint dataset.

Value

A vector of consensus cell type annotations, merging inputs from the three datasets.

Note

This function is designed to harmonize specific cell types, especially T cells, B cells, monocytic cells, and innate lymphoid cells (ILCs), across reference datasets.

See also

Examples

# Example usage:
tibble::tibble(
  azimuth_predicted.celltype.l2 = c("CD8 TEM", "NK", "CD4 Naive"),
  monaco_first.labels.fine = c("Effector memory CD8 T cells", "Natural killer cells", "Naive CD4 T cells"),
  blueprint_first.labels.fine = c("CD8+ Tem", "NK cells", "Naive B-cells")
) |>
  dplyr::mutate(consensus = reference_annotation_to_consensus(
    azimuth_predicted.celltype.l2, monaco_first.labels.fine, blueprint_first.labels.fine))
#> # A tibble: 3 × 4
#>   azimuth_predicted.ce…¹ monaco_first.labels.…² blueprint_first.labe…³ consensus
#>   <chr>                  <chr>                  <chr>                  <glue>   
#> 1 CD8 TEM                Effector memory CD8 T… CD8+ Tem               cd8 tem  
#> 2 NK                     Natural killer cells   NK cells               nk       
#> 3 CD4 Naive              Naive CD4 T cells      Naive B-cells          cd4 naive
#> # ℹ abbreviated names: ¹​azimuth_predicted.celltype.l2,
#> #   ²​monaco_first.labels.fine, ³​blueprint_first.labels.fine