Apply a transformation to an assay and save as HDF5

This function applies a specified transformation to the assay of a SummarizedExperiment object and saves the transformed object in HDF5 format.

transform_utility(
  input_read_RNA_assay,
  transform_fx,
  external_path,
  container_type,
  scale_max
)

Arguments

input_read_RNA_assay

A SummarizedExperiment object to be transformed.

transform_fx

A function to apply to the assay of the SummarizedExperiment object.

external_path

A character string specifying the directory path to save the transformed object.

container_type

A character vector specifying the output file type. Ideally it should match to the input file type.

scale_max

An integer specifying the allowed max count value. It is used to define scaling factor in the process of transformation.

Value

The function does not return an object. It saves the transformed SummarizedExperiment object to the specified path.